Split page.tsx into two pages to fix prerender error
Added a page called recipesearchpage.tsx so when the page search function gets called it is wrapped in a suspense tag
